move update parsed record to service
