+++ mod-source-record-storage-server/src/main/java/org/folio/services/impl/ParsedRecordServiceImpl.java
@@ -5,9 +5,13 @@ import java.util.Objects;
 import java.util.stream.Collectors;
 
 import javax.ws.rs.BadRequestException;
+import javax.ws.rs.NotFoundException;
 
+import org.folio.dao.LBRecordDao;
 import org.folio.dao.ParsedRecordDao;
 import org.folio.dao.query.ParsedRecordQuery;
+import org.folio.rest.jaxrs.model.ExternalIdsHolder;
+import org.folio.rest.jaxrs.model.Metadata;
 import org.folio.rest.jaxrs.model.ParsedRecord;
 import org.folio.rest.jaxrs.model.ParsedRecordCollection;
 import org.folio.rest.jaxrs.model.ParsedRecordsBatchResponse;
@@ -15,6 +19,7 @@ import org.folio.rest.jaxrs.model.Record;
 import org.folio.rest.jaxrs.model.RecordCollection;
 import org.folio.services.AbstractEntityService;
 import org.folio.services.ParsedRecordService;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import io.vertx.core.CompositeFuture;
@@ -24,11 +29,14 @@ import io.vertx.core.Promise;
 @Service
 public class ParsedRecordServiceImpl extends AbstractEntityService<ParsedRecord, ParsedRecordCollection, ParsedRecordQuery, ParsedRecordDao> implements ParsedRecordService {
 
+  @Autowired
+  private LBRecordDao recordDao;
+
   @Override
   public Future<ParsedRecordsBatchResponse> updateParsedRecords(RecordCollection recordCollection, String tenantId) {
     List<Future> futures = recordCollection.getRecords().stream()
       .map(this::validateParsedRecordId)
-      .map(record -> dao.update(record.getParsedRecord(), tenantId))
+      .map(record -> updateParsedRecord(record, tenantId))
       .collect(Collectors.toList());
     Promise<ParsedRecordsBatchResponse> promise = Promise.promise();
     CompositeFuture.join(futures).onComplete(ar -> {
@@ -44,6 +52,19 @@ public class ParsedRecordServiceImpl extends AbstractEntityService<ParsedRecord,
       promise.complete(response);
     });
     return promise.future();
+  }
+
+  @Override
+  public Future<ParsedRecord> updateParsedRecord(Record record, String tenantId) {
+    String id = record.getId();
+    ParsedRecord parsedRecord = record.getParsedRecord();
+    ExternalIdsHolder externalIdsHolder = record.getExternalIdsHolder();
+    Metadata metadata = record.getMetadata();
+    return dao.inTransaction(tenantId, connection ->
+      recordDao.getById(connection, id, tenantId)
+        .map(r -> r.orElseThrow(() -> new NotFoundException(String.format("Couldn't find record with id %s", id))))
+        .compose(r -> recordDao.save(connection, r.withExternalIdsHolder(externalIdsHolder).withMetadata(metadata), tenantId))
+        .compose(r -> dao.save(connection, parsedRecord, tenantId)));
   }
 
   private Record validateParsedRecordId(Record record) {
