+++ mod-source-record-storage-server/src/test/java/org/folio/services/impl/ParsedRecordServiceTest.java
@@ -2,6 +2,7 @@ package org.folio.services.impl;
 
 import org.apache.commons.lang3.reflect.FieldUtils;
 import org.folio.ParsedRecordMocks;
+import org.folio.dao.LBRecordDao;
 import org.folio.dao.ParsedRecordDao;
 import org.folio.dao.query.ParsedRecordQuery;
 import org.folio.rest.jaxrs.model.ParsedRecord;
@@ -16,11 +17,15 @@ import io.vertx.ext.unit.junit.VertxUnitRunner;
 @RunWith(VertxUnitRunner.class)
 public class ParsedRecordServiceTest extends AbstractEntityServiceTest<ParsedRecord, ParsedRecordCollection, ParsedRecordQuery, ParsedRecordDao, ParsedRecordService, ParsedRecordMocks> {
 
+  LBRecordDao mockRecordDao;
+
   @Override
   public void createService(TestContext context) throws IllegalAccessException {
     mockDao = Mockito.mock(ParsedRecordDao.class);
+    mockRecordDao = Mockito.mock(LBRecordDao.class);
     service = new ParsedRecordServiceImpl();
     FieldUtils.writeField(service, "dao", mockDao, true);
+    FieldUtils.writeField(service, "recordDao", mockRecordDao, true);
   }
 
   @Override
